# concurrently load weights of DeepseekV2ForCausalLM