<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

The previous patch https://github.com/sgl-project/sglang/pull/7277 supported multi-thread model weight loading (from NVMe to CPU memory), significantly reducing the model weight loading time by fully utilizing the NVMe bandwidth. Building on this, the current patch modifies the load_weight() function of DeepseekV2ForCausalLM to enable multi-threaded concurrent loading of weights (from CPU memory to device memory), further shortening the loading time.

## Modifications

<!-- Describe the changes made in this PR. -->

Introduce `concurrent.futures.ThreadPoolExecutor` and change the invocation of weight_loader from serial to parallel.

Test DeepSeek-R1 safetensor loading in NVME ssd PCIE Gen4 (Speed 16GT/s, Width x4)

Server args
```
python -m sglang.launch_server \
  --model-path /data/DeepSeek-R1 \
  --tp-size 8 \
  --max-running-requests 4 \
  --mem-fraction-static 0.9 \
  --model-loader-extra-config '{"enable_multithread_load": true, "num_threads": 8}'
```

> OPT-1: https://github.com/sgl-project/sglang/pull/7277
> OPT-2: this patch

|      | Time consumption before optimization | Time consumption with OPT-1 |  Time consumption with OPT-1 && OPT-2 | 
| ----------- | ----------- | ---- | --- |
| Weights in LVM striped logical volume with 3 NVME ssd |  217s     | 115s | 65s |
| Weights in pagecache (cpu memory) | 60s  | 56s | 32s |

## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.